# openapi_dump.json
{"openapi":"3.1.0","info":{"title":"MSIS API","version":"0.0.1"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/search":{"get":{"summary":"Search","description":"搜索公文文档","operationId":"search_api_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":1,"description":"搜索关键词","title":"Q"},"description":"搜索关键词"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/templates":{"get":{"summary":"List Templates","description":"获取模板列表","operationId":"list_templates_api_templates_get","parameters":[{"name":"type","in":"query","required":false,"schema":{"type":"string","title":"Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TemplatesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Template","description":"创建模板","operationId":"create_template_api_templates_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTemplateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/templates/{template_id}/render":{"post":{"summary":"Render Template","description":"渲染模板","operationId":"render_template_api_templates__template_id__render_post","parameters":[{"name":"template_id","in":"path","required":true,"schema":{"type":"integer","title":"Template Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RenderRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/config":{"get":{"summary":"Get Document Config","description":"获取文档库配置","operationId":"get_document_config_api_documents_config_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigResponse"}}}}}},"put":{"summary":"Update Document Config","description":"更新文档库配置","operationId":"update_document_config_api_documents_config_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateConfigRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/upload":{"post":{"summary":"Upload Document","description":"上传文档到文档库","operationId":"upload_document_api_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_documents_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/list":{"get":{"summary":"List Documents","description":"获取已索引的文档列表","operationId":"list_documents_api_documents_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentsListResponse"}}}}}}},"/api/documents/reindex":{"post":{"summary":"Reindex Documents","description":"手动触发重新索引","operationId":"reindex_documents_api_documents_reindex_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReindexResponse"}}}}}}},"/api/documents/{file_name}":{"delete":{"summary":"Delete Document","description":"删除已索引的文档","operationId":"delete_document_api_documents__file_name__delete","parameters":[{"name":"file_name","in":"path","required":true,"schema":{"type":"string","title":"File Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_document_api_documents_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_document_api_documents_upload_post"},"ConfigResponse":{"properties":{"document_path":{"type":"string","title":"Document Path"},"index_path":{"type":"string","title":"Index Path"},"auto_reindex":{"type":"boolean","title":"Auto Reindex"},"supported_formats":{"items":{"type":"string"},"type":"array","title":"Supported Formats"}},"type":"object","required":["document_path","index_path","auto_reindex","supported_formats"],"title":"ConfigResponse"},"CreateTemplateRequest":{"properties":{"name":{"type":"string","title":"Name"},"type":{"type":"string","title":"Type"},"description":{"type":"string","title":"Description","default":""}},"type":"object","required":["name","type"],"title":"CreateTemplateRequest"},"DocumentItem":{"properties":{"name":{"type":"string","title":"Name"},"size":{"type":"integer","title":"Size"},"path":{"type":"string","title":"Path"}},"type":"object","required":["name","size","path"],"title":"DocumentItem"},"DocumentsListResponse":{"properties":{"documents":{"items":{"$ref":"#/components/schemas/DocumentItem"},"type":"array","title":"Documents"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["documents","total"],"title":"DocumentsListResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ReindexResponse":{"properties":{"total":{"type":"integer","title":"Total"},"success":{"type":"integer","title":"Success"},"errors":{"items":{"type":"string"},"type":"array","title":"Errors","default":[]}},"type":"object","required":["total","success"],"title":"ReindexResponse"},"RenderRequest":{"properties":{"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"}},"type":"object","required":["title","content"],"title":"RenderRequest"},"SearchResponse":{"properties":{"query":{"type":"string","title":"Query"},"results":{"items":{},"type":"array","title":"Results"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["query","results","total"],"title":"SearchResponse"},"TemplatesResponse":{"properties":{"templates":{"items":{},"type":"array","title":"Templates"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["templates","total"],"title":"TemplatesResponse"},"UpdateConfigRequest":{"properties":{"document_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Path"},"auto_reindex":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Auto Reindex"}},"type":"object","title":"UpdateConfigRequest"},"UploadResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"file_name":{"type":"string","title":"File Name"},"message":{"type":"string","title":"Message"},"indexed_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Indexed Path"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["success","file_name","message"],"title":"UploadResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}